+++ clang/docs/ReleaseNotes.rst
@@ -129,6 +129,15 @@ Improvements to Clang's diagnostics
 - ``-Wunused-variable`` no longer warn for references extending the lifetime
   of temporaries with side effects. This fixes `Issue 54489
   <https://github.com/llvm/llvm-project/issues/54489>`_.
+- Modified the behavior of ``-Wstrict-prototypes`` and added a new, related
+  diagnostic ``-Wdeprecated-non-prototype``. The strict prototypes warning will
+  now only diagnose deprecated declarations and definitions of functions
+  without a prototype where the behavior in C2x will remain correct. This
+  diagnostic remains off by default but is now enabled via ``-pedantic`` due to
+  it being a deprecation warning. ``-Wdeprecated-non-prototype`` will diagnose
+  cases where the deprecated declarations or definitions of a function without
+  a prototype will change behavior in C2x. This diagnostic is grouped under the
+  ``-Wstrict-prototypes`` warning group, but is enabled by default.
 
 Non-comprehensive list of changes in this release
 -------------------------------------------------